toolkit/components/passwordmgr/test/mochitest/test_username_focus.html


<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test interaction between autocomplete and focus on username fields</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script>
 const readyPromise = addLoginsInParent(
    ["http://mochi.test:8888", "http://username-focus-1", null, "testuser1A", "testpass1A", "", ""],
    ["http://mochi.test:8888", "http://username-focus-2", null, "testuser2A", "testpass2A", "", ""],
    ["http://mochi.test:8888", "http://username-focus-2", null, "testuser2B", "testpass2B", "", ""]
  ).then(() => registerRunTests(5));
</script>

<p id="display"></p>
<div id="content">
  <!-- first 3 forms have a matching user+pass login -->

  <!-- user+pass form. -->
  <form id="form-autofilled" action="http://username-focus-1">
    <input  type="text"     name="uname">
    <input  type="password" name="pword">
    <button type="submit" name="submit">Submit</button>
  </form>

  <!-- user+pass form, username prefilled -->
  <form id="form-autofilled-prefilled-un" action="http://username-focus-1">
    <input  type="text"     name="uname" value="testuser1A">
    <input  type="password" name="pword">
    <button type="submit">Submit</button>
  </form>

  <!-- user+pass form. -->
  <form id="form-autofilled-focused-dynamic" action="http://username-focus-1">
    <input  type="text"             name="uname">
    <input  type="not-yet-password" name="pword">
    <button type="submit">Submit</button>
  </form>


  <!-- next 5 forms have matching user+pass (2x) logins -->

  <!-- user+pass form. -->
  <form id="form-multiple" action="http://username-focus-2">
    <input  type="text"     name="uname">
    <input  type="password" name="pword">
    <button type="submit">Submit</button>
  </form>

  <!-- user+pass form dynamic with existing focus -->
  <form id="form-multiple-dynamic" action="http://username-focus-2">
    <input  type="text"             name="uname">
    <input  type="not-yet-password" name="pword">
    <button type="submit">Submit</button>
  </form>

  <!-- user+pass form, username prefilled -->
  <form id="form-multiple-prefilled-un1" action="http://username-focus-2">
    <input  type="text"     name="uname" value="testuser2A">
    <input  type="password" name="pword">
    <button type="submit">Submit</button>
  </form>

  <!-- user+pass form, different username prefilled -->
  <form id="form-multiple-prefilled-un2" action="http://username-focus-2">
    <input  type="text"     name="uname" value="testuser2B">
    <input  type="password" name="pword">
    <button type="submit">Submit</button>
  </form>

  <!-- user+pass form, username prefilled with existing focus -->
  <form id="form-multiple-prefilled-focused-dynamic" action="http://username-focus-2">
    <input  type="text"             name="uname" value="testuser2B">
    <input  type="not-yet-password" name="pword">
    <button type="submit">Submit</button>
  </form>

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
function removeFocus() {
  getFormElementByName("-autofilled", "submit").focus();
}

add_setup(async () => {
  ok(readyPromise, "check promise is available");
  await readyPromise;
});

add_task(async function test_autofilled() {
  const usernameField = getFormElementByName("-autofilled", "uname");
  info("Username and password already filled so don't show autocomplete");
  await noPopupBy(() => usernameField.focus());

  removeFocus();
  usernameField.value = "testuser";
  info("Focus when we don't have an exact match");
  await popupBy(() => usernameField.focus());
});

add_task(async function test_autofilled_prefilled_un() {
  const usernameField = getFormElementByName("-autofilled-prefilled-un", "uname");
  info("Username and password already filled so don't show autocomplete");
  await noPopupBy(() => usernameField.focus());

  removeFocus();
  usernameField.value = "testuser";
  info("Focus when we don't have an exact match");
  await popupBy(() => usernameField.focus());
});

add_task(async function test_autofilled_focused_dynamic() {
  const usernameField = getFormElementByName("-autofilled-focused-dynamic", "uname");
  const passwordField = getFormElementByName("-autofilled-focused-dynamic", "pword");
  info("Username and password will be filled while username focused");
  await noPopupBy(() => usernameField.focus());
  info("triggering autofill");
  await noPopupBy(() => passwordField.type = "password");

  const popupState = await getPopupState();
  is(popupState.open, false, "Check popup is closed");

  removeFocus();
  passwordField.value = "test";
  info("Focus when we don't have an exact match");
  await popupBy(() => usernameField.focus());
});

// Begin testing forms that have multiple saved logins

add_task(async function test_multiple() {
  const usernameField = getFormElementByName("-multiple", "uname");
  info("Fields not filled due to multiple so autocomplete upon focus");
  await popupBy(() => usernameField.focus());
});

add_task(async function test_multiple_dynamic() {
  const usernameField = getFormElementByName("-multiple-dynamic", "uname");
  const passwordField = getFormElementByName("-multiple-dynamic", "pword");
  info("Fields not filled but username is focused upon marking so open");
  await noPopupBy(() => usernameField.focus());

  info("triggering _fillForm code");
  await popupBy(() => passwordField.type = "password");
});

add_task(async function test_multiple_prefilled_un1() {
  const usernameField = getFormElementByName("-multiple-prefilled-un1", "uname");
  info("Username and password already filled so don't show autocomplete");
  await noPopupBy(() => usernameField.focus());

  removeFocus();
  usernameField.value = "testuser";
  info("Focus when we don't have an exact match");
  await popupBy(() => usernameField.focus());
});

add_task(async function test_multiple_prefilled_un2() {
  const usernameField = getFormElementByName("-multiple-prefilled-un2", "uname");
  info("Username and password already filled so don't show autocomplete");
  await noPopupBy(() => usernameField.focus());

  removeFocus();
  usernameField.value = "testuser";
  info("Focus when we don't have an exact match");
  await popupBy(() => usernameField.focus());
});

add_task(async function test_multiple_prefilled_focused_dynamic() {
  const usernameField = getFormElementByName("-multiple-prefilled-focused-dynamic", "uname");
  const passwordField = getFormElementByName("-multiple-prefilled-focused-dynamic", "pword");
  info("Username and password will be filled while username focused");
  await noPopupBy(() => usernameField.focus());
  info("triggering autofill");
  await noPopupBy(() => passwordField.type = "password");

  const popupState = await getPopupState();
  is(popupState.open, false, "Check popup is closed");

  removeFocus();
  passwordField.value = "test";
  info("Focus when we don't have an exact match");
  await popupBy(() => usernameField.focus());
});
</script>
</pre>
</body>
</html>
